requirements.txt
django>=4.2
djangorestframework>=3.12
django-filter>=22.1
django-cors-headers>=3.13
drf-spectacular>=0.27
django-redis>=5.2
redis>=4.0
gunicorn